refactor gameLoop into React Hook & use single component for web + native
