+++ app/src/App/index.test.tsx
@@ -2,6 +2,8 @@ import "react-native";
 import React from "react";
 import App from "./index";
 
+jest.useFakeTimers();
+
 // Note: test renderer must be required after react-native.
 import renderer from "react-test-renderer";
